move path to the argument settings
